[AN] refactor(#603): 회원가입 시 깃허브 아이디 별도로 입력 받지 않도록 수정
